Add fixes, refactors, and formatting from writing tests
This commit contains a number of fixes, refactors, and reformats found during the writing of unit tests for this repo.  These fixes are what is needed to take the previous code (which was essentially very detailed pseudocode) and implement the actual features.  Tests are committed separately to make things a little better isolated.