feat: add guard in CharmReconciler.add() to prevent calling after .install()
CharmReconciler.install() needs the entire list of Components in order to subscribe the reconciler to any custom events.  If we do:

reconciler.install()
reconciler.add()

The added charm's custom events will not be observed.  This commit adds a check to raise an Exception when this is done.

A more advanced fix to this could be always re-running .install() when .add() is called (or at least when we .add() with a new custom event).  Not sure if the framework de-duplicates event subscriptions, so we might need to guard against requesting the same event handling multiple times.