feat: use ops's new collect-status event for final status handling
This uses the collect-status event, implemented in ops 2.5, to set the charm's status on exit of any events.  This sits alongside the regular status pool - the status pool is still used internally in the charm, but the final status is set through the collect-status mechanism.

This is a prototype and may need more thought/testing, but after initial testing of config-changed and update-status events seems like it works great!