remove UnknownStatus from status prioritizer
The status prioritizer previously would report an UnknownStatus when
it had no components to report statuses.  UnknownStatus is a readonly status
in Juju so this is not appropriate.  This commit changes Prioritizer to
identify when it has no components and instead return ActiveStatus (since a
Prioritizer without anything is effectively Active).