fix(tests): Fetch alert rules dynamically (#140)
Refactor integration tests to fetch alert rules dynamically. This ensures that alert rules being checked are up-to-date and simulate a real-world scenario, since assert_alert_rules is meant to be used in charm repositories. This is achieved by:

Downloading the .charm artifact
Extracting it
Using get_alert_rules() on the extracted path.
Addresses part 1 of #139