feat: apply ambient charms sequentially (#177)
This PR was inspired by the flakiness we've seen in a lot of PRs (i.e. https://github.com/canonical/envoy-operator/pull/188, https://github.com/canonical/kubeflow-tensorboards-operator/pull/221) in which the tests were failing on the `build_and_deploy` step.

A lot of the times the culprit was the Istio charms, in which they were getting to error state and then taking a lot of time to recover.

This PR aims to resolve this by installing, and waiting for idle, the Istio ambient charms **sequentially**. So:
1. First the `istio-k8s` charm should be active
2. Then the `istio-ingress-k8s`
3. Then the `istio-beacon-k8s`, which also adds the model to the namespace